fix(settings): emails actions a11y and design
